Can display all users from GUI

+++ master/working/src/controllers/UserController.java
@@ -2,12 +2,24 @@ package controllers;
 
 import domain.User;
 import domain.UserType;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
+import javafx.event.EventHandler;
 import javafx.scene.control.*;
+import javafx.scene.control.cell.PropertyValueFactory;
+import javafx.scene.control.cell.TextFieldTableCell;
 import javafx.stage.Stage;
+import javafx.util.StringConverter;
+import javafx.util.converter.BooleanStringConverter;
+import javafx.util.converter.DateStringConverter;
+import javafx.util.converter.DoubleStringConverter;
+import javafx.util.converter.IntegerStringConverter;
 import logic.AuthenticationSystem;
 
 
 import javax.xml.soap.Text;
+import java.util.Date;
+import java.util.List;
 
 /**
  * Created by DillonVaghela on 2/20/17.
@@ -22,6 +34,13 @@ public class UserController {
     public ComboBox UT;
     public TextField SUID;
     private AuthenticationSystem auth = AuthenticationSystem.getInstance();
+    public TableView<User> tUsers;
+    public TableColumn<User, String> userID;
+    public TableColumn<User, String> password;
+    public TableColumn<User, String> firstname;
+    public TableColumn<User, String> surname;
+    public TableColumn<User, UserType> userType;
+    final ObservableList<User> tableEntries = FXCollections.observableArrayList();
 
     public void addEditUser() throws Exception
     {
@@ -53,7 +72,7 @@ public class UserController {
                     userType = UserType.NORMAL;
                 }
                 boolean checker = auth.addEditUser(UID.getText(), P.getText(), FN.getText(),SN.getText(),userType);
-                showAlert("User added: " + checker);
+                showAlert("User "+addOrEdit+": " + checker);
                 if (checker)
                 {
                     Stage addStage = (Stage) UID.getScene().getWindow();
@@ -72,6 +91,10 @@ public class UserController {
                     return;
                 }
                 else {
+                    if (!showAlertC("Sure you want to delete?"))
+                    {
+                        return;
+                    }
                     boolean check = auth.deleteUser(UID.getText());
                     showAlert("User deleted: " + check);
                 }
@@ -98,7 +121,24 @@ public class UserController {
 
     public void setUserDets(User user)
     {
-        
+        UID.setText(user.getUserID());
+        UID.setDisable(false);
+        P.setText(user.getPassword());
+        P.setDisable(false);
+        FN.setText(user.getFirstName());
+        FN.setDisable(false);
+        SN.setText(user.getSurname());
+        SN.setDisable(false);
+        String theUser;
+         if (user.getUserType().toString().equals("NORMAL"))
+         {
+             theUser = "Normal";
+         }
+         else {
+             theUser = "Admin";
+         }
+        UT.setValue(theUser);
+        UT.setDisable(false);
     }
 
     public boolean checkFields()
@@ -117,6 +157,12 @@ public class UserController {
         alert.showAndWait();
     }
 
+    public void AllUsers()
+    {
+        List<User> arrayList = auth.getUsersList();
+        DisplayTable(arrayList);
+    }
+
     public boolean showAlertC(String message)
     {
         Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
@@ -128,6 +174,86 @@ public class UserController {
         } else {
             return false;
         }
+    }
+
+    public void DisplayTable(List<User> arrayList)
+    {
+        try {
+            tUsers.setDisable(false);
+            tableEntries.removeAll(tableEntries);
+            for(int i =0; i<arrayList.size(); i++){
+
+                tableEntries.add(arrayList.get(i));
+            }
+
+            userID.setCellValueFactory(new PropertyValueFactory<User, String>("userID"));
+            userID.setCellFactory(TextFieldTableCell.<User>forTableColumn( ));
+            //userID.setOnEditCommit(new EventHandler<TableColumn.CellEditEvent<User, String>>() {
+             //   @Override
+              //  public void handle(TableColumn.CellEditEvent<User, String> event) {
+               //     ( event.getTableView().getItems().get(event.getTablePosition().getRow())).set(event.getNewValue());
+                //}
+            //});
+
+            password.setCellValueFactory(new PropertyValueFactory<User, String>("password"));
+            password.setCellFactory(TextFieldTableCell.<User>forTableColumn());
+            password.setOnEditCommit(new EventHandler<TableColumn.CellEditEvent<User, String>>() {
+                @Override
+                public void handle(TableColumn.CellEditEvent<User, String> event) {
+                    ( event.getTableView().getItems().get(event.getTablePosition().getRow())).setPassword(event.getNewValue());
+                }
+            });
+
+            firstname.setCellValueFactory(new PropertyValueFactory<User, String>("firstName"));
+            firstname.setCellFactory(TextFieldTableCell.<User>forTableColumn( ));
+            firstname.setOnEditCommit(new EventHandler<TableColumn.CellEditEvent<User, String>>() {
+                @Override
+                public void handle(TableColumn.CellEditEvent<User, String> event) {
+                    ( event.getTableView().getItems().get(event.getTablePosition().getRow())).setFirstName(event.getNewValue());
+                }
+            });
+
+            surname.setCellValueFactory(new PropertyValueFactory<User, String>("surname"));
+            surname.setCellFactory(TextFieldTableCell.<User>forTableColumn());
+            surname.setOnEditCommit(new EventHandler<TableColumn.CellEditEvent<User, String>>() {
+                @Override
+                public void handle(TableColumn.CellEditEvent<User, String> event) {
+                    ( event.getTableView().getItems().get(event.getTablePosition().getRow())).setSurname(event.getNewValue());
+                }
+            });
+
+
+            userType.setCellValueFactory(new PropertyValueFactory<User, UserType>("userType"));
+            userType.setCellFactory(TextFieldTableCell.<User, UserType>forTableColumn(new StringConverter<UserType>() {
+                @Override
+                public String toString(UserType object) {
+                    return null;
+                }
+
+                @Override
+                public UserType fromString(String string) {
+                    return null;
+                }
+            }));
+            userType.setOnEditCommit(new EventHandler<TableColumn.CellEditEvent<User, UserType>>() {
+                @Override
+                public void handle(TableColumn.CellEditEvent<User, UserType> event) {
+
+                    ( event.getTableView().getItems().get(event.getTablePosition().getRow())).setUserType(event.getNewValue());
+                }
+            });
+
+
+
+            tUsers.setItems(tableEntries);
+        }
+        catch (Exception e)
+        {
+            System.out.println("not working");
+            e.printStackTrace(  );
+            System.out.println(e);
+        }
+
     }
 
 }
